Test circular data stack with push & pop
This uses SP for the return stack and adds a new circular data
stack built from registers R12..R15 and 16 quad-words of memory.
The Top and Second elements of the stack live in R12 and R13. The
rest of the items on the stack get stored in a 16 qword circular
buffer with R14 as the head index and R15 as the length.

The main benefits of a small-ish circular stack are:
1. It can't segfault. You may get logic errors for overflow or
   underflow conditions, but those are easier to deal with.
2. Small stack encourages good habits with code factoring and
   memory management. Data structures of any complexity should go
   in memory and be passed by reference on the stack.