Use System V ABI stack from exec as return stack
Previously, I tried changing the value of SP to make my own return
stack. This time, I went through the System V ABI AMD64 supplement
and learned how to use the stack provided for my process by exec.

stacktest2.nasm dumps the initial stack frame with argc, argv, and
envp. But, it's not yet smart enough to follow the pointers in
argv and envp to print their strings.

Also, I learned that doing `lea rxx, [ryy + 8 * rzz]` style
indirect addressing will stomp on the contents of RAX and RDX. So,
it's important to assign registers carefully in that context.