Add token threaded inner interpreter
This was based on a copy of string1.nasm which used a bunch of
macros and CALLs to fake up a sequence of interpreter calls.

Now, with token1.nasm, the macros are mostly gone, some of the
CALLs are now tail call JMPs, and there's actually a token
interpreter loop that executes a hand-compiled list of virtual
machine instructions.