Improve makefile rules for objdump disassembly
My old way of invoking ld did not preserve the full symbol table
from nasm. This change makes it so the labels from nasm show up in
the objdump disassembly. I also turned on the option to visually
mark loops. The point is to make it easier to experiment with ideas
for reducing code size.