Experiment with non-line-buffered terminal input
This is a C program using the POSIX functions in termios.h to set
up character buffered keyboard input for use with the read()
syscall. Goal is to get moderately raw keybard input where the host
OS handles keyscan to character conversion, but the terminal does
not lock up for ^S and ^Q XON/XOFF flow control, and it does not
send POSIX signals for ^C, ^D, ^Z, or ^\.

The idea is to pass through emacs-style control and meta keyboard
shortcuts so the REPL can use them to navigate and make edits in
its input buffer.

I might try to use this C code as a wrapper for my Forth outer
interpreter written in assembly. Or, I might do the termios stuff
in assembly. Using a C wrapper would have the advantage of easier
POSIX compatibility, perhaps yielding a slower bitrot rate as
kernel syscall and ioctl stuff changes over time.