Fix long input line scrolling behavior
This makes the terminal output look more reasonable when typing an
input line that is wider than the terminal window. Previously, the
input buffer painting code would would pollute the scrollback
buffer with many partial wrapped lines, due to my erase-current-row
redrawing method not being smart enough to detect wrapping.

This new code checks the terminal row and column size, then
adjusts the line redrawing so that only the last n characters of
the input buffer get drawn, where n is the width of the terminal
in columns.