Make repl backspace work on grapheme clusters
This makes the REPL backspace action aware of UTF-8 encoding and
some of the joining and modifying codepoints that occur in grapheme
clusters for modern emoji. The point is to make backspace operate
in units of user percieved characters (grapheme clusters), which
may be one, some, or many bytes in the terminal input buffer.