Switch bye flag from memory to register
For the moment, I'm using a flag to mark that `bye` has been
invoked. This could be implemented more simply with an `exit()`
syscall, but using the flag allows for more of an orderly shutdown
by way of the normal state machines. I'm expecting this will be
useful for things like avoiding audio glitches as the VM shuts
down.