Add dictionary lookup to outer interpreter
This adds a string search routine to the outer interpreter so that
it can now resolve words like `.S`, `Nop`, and `Bye` to tokens and
have the inner interpreter run them. I haven't implemented anything
yet for parsing numbers, so the outer interpreter isn't very useful
yet. But, it's close.

Next up: attempt to parse words that are not found in the
dictionary into numbers and push the numbers onto the stack.