Overhaul outer interpreter to get ." working
This refactors the outer interpreter to use an index variable (IN)
to track the next available input byte in the terminal input
buffer. With that mechanism in place, it becomes possible to start
letting token handler routines consume bytes from the input stream.
In particular, the ." (dot-quote) word needs to be able to do that.