Add number pushing to outer interpreter
This adds decimal and hex number conversions to the outer
interpreter. When a word is not found in the dictionary, the
interpreter will now attempt to convert it to a number in the
current base (decimal or hex) and push that number to the stack.

At this point, basic integer math stack operations seem to be
working pretty well. Next step is to add defining words for adding
new words to the dictionary at runtime.