Make number formatting a lot smarter
This makes the `.` and `.S` number formatting words smarter so they
can use the number base to format 32-bit numbers as signed decimal
or unsigned hexadecimal.

Also, `.S` now formats the stack horizontal instead of vertical.
This matches how other Forth systems seem to work, but it requires
you to know that the bottom stack cell is on the left and the top
stack cell is on the right.