Fix negative number parsing bug
This is wild. I was checking the wrong register (dl instead of r9b)
to look for a leading '-' on number words. The crazy thing is that
it actually worked for -1 down to -999! But, as soon as I tried to
do -1000 or less, it stopped working. This seems to be the result
of using the dl register in my dictionary name matching, perhaps
with the word length check behaving differently for strings up to
length 4 bytes, compared to 5 bytes or more.

Anyhow, I'm testing the correct register now, and it seems to work.