Fix bug in error message for bad token
The code to display an error message when the inner interpreter
encounters a bad token was being silly about handling the pointer
to the offending token.

After this change, the instruction pointer in the error message
should match with the address in a hexdump of the .data section in
the linked binary.

I tested the error message by inserting a `db 99` before the final
next token in the loadscreen area. But, I removed that line prior
to making this commit.