Rework loadscreen (embed Forth source from file)
My old "loadscreen" code was an array of hand-compiled token
constants written as nasm `db` lines for the .data section. This
switches to using the `incbin` macro to include Forth source code
from the `screen00.fs` file in .data as a binary blob.

This commit also modifies the outer interpreter to use CR and LF as
word boundary characters, in addition to space. So, the outer
interpreter can now take multi-line strings as input.