Fix bug in handling of 4-byte UTF-8 sequences
My terminal IO code was using a faulty method of detecting when a
read() call got interrupted by a SIGWINCH signal. I was looking for
`result == EINTR`, assuming that EINTR was negative. But, its value
value was actually `4`. So, any time that read()'s result was 4
bytes, that input was getting discarded. The effect was to filter
out many emoji characters that encode as 4 UTF-8 bytes.