Refactor word scanning
This refactors the code for scanning the input stream for the next
input word. Now the outer interpreter and the CREATE word can share
code from the WORD word which copies a word from the input stream
to the scratch area after the last dictionary entry.

There is also a safety check in CREATE so that it will always leave
at least 256 bytes in the dictionary memory. The point of that is
to make sure WORD can be used by the outer loop to accept input
from the shell, even when the dictionary is "full".