Add compiling for number literals
This allows number literals to be included in colon definitions so
they will cause the push to happen when the word is run, rather
than when the word is defined.

There are five different tokens used for literals: U8, I8, U16,
I16, and I32. The number compiler picks a token based on the number
of bytes needed to accurately represent the number. I also added
tests to cover the different sizes and signs of number literals.