Implement jump token with tail call optimization
These changes allow code like the `a ;` in this example,
```
: a ... ;  : b ... a ;
```
to optimize with a jump to `a` at the end of `b` rather than a call
followed by a next. This allows for tail recursion without filling
up the return stack.

There's also a loop limit counter that constrains the total number
of VM instruction tokens that can be used to implement one word as
invoked from the outer interpreter (command line shell). This is to
prevent runaway loops from locking up the whole interpreter, since
it is single-threaded.

I also added some tests to demonstrate how the tail call optimizer
and loop limit mechanism work.