Rename "Next" VM instruction token as "Return"
My naming for VM token opcodes has become a problem as I work on
extending my compiler for conditional jumps and looping constructs.
It's getting really confusing.

This is the start of a refactor to bring my usage closer to the
instruction sets and vocabularies in the lineage of Chuck Moore's
non-ANS Forths (machineForth, arrayForth, colorForth,...).

Also, where the new fancy stuff gets weird or confusing, I'm looking
back to figForth, Forth-79, and ANS for what appears to be the least
confusing option for stuff like >R, R>, INVERT, 0=, and NEGATE.