Switch stack depth registers (r14/15 -> xmm0/1)
This frees up registers r14 and r15, which I was using to track
stack depth, by switching to the SSE registers xmm0 and xmm1. R14
and r15 are special because the amd64 linux ABI specifies them as
preserved by the callee. So, basically, I can use them to preserve
state across regular calls, and syscalls, without troubles.

Immediate goal of shuffling registers is to make room for adding a
VM register `R` (top of return stack) to use with the `Next` word
for loop indexing. Might also add a `P` VM register for the program
counter.