Replace `not` VM op with `invert` and `negate`
This is a very old Forth problem where the original figForth and
Forth-79 used NOT to mean a one's complement bitwise inversion
that also happened to double as a boolean not.

Apparently, it was used for doing things like `NOT IF ...` using
NOT as a boolean operation on 0 (false) or -1 (true), as was the
old Forth convention for truth values. But, that usage was
ambiguous and confusing.

As best I can make out, by the time of either Forth-83 or ANS, the
problem was resolved by replacing NOT with the INVERT, NEGATE, and
0= words. This helps avoid confusion about whether the intended
operation is one's complement, two's complement, or boolean.