Partially fix save/restore of SSE registers
I started using the XMM0 and XMM1 SSE registers for persistent
state in the Forth VM. But, that's kinda problematic because those
are supposed to be saved by the caller (AMD64 System V ABI says so),
and I can't do that yet because of how I factored my terminal input
state machine.

So, this commit does the proper push and pop thing in the one spot
where I can. The other spot gets a big fat TODO comment explaining
how it works fine now, but adding floating point calls could make
a big mess.

I'm hoping that once I get the compiler filled in better, I can
refactor a lot of stuff to make this simpler and less precarious.