Fix terminal input + save SSE registers properly
Previously, I was calling back and forth between my C code and my
libmarkab assembly code in a way that made it impossible to
properly save and restore the xmm0 and xmm1 SSE registers across
all of the C<-->assembly function calls.

This commit untangles that mess and does what I hope are all of the
proper ABI-compliant incantations to save registers and align the
stack.