Add register R as top of return stack
Previously, all items of the return stack were stored in memory.
Now, the top item, R, is stored in a register so it will be faster
to use it as a loop counter.

This commit also cleans up a few spots I missed when switching from
my old usage of "Next" to what I now call "Return".