Add fizzbuzz test
This is a joke, kinda? I wanted to see if I could build the
equivalent of an if statement from only my `next` conditional
return word. It works, but the contortions I had to go through to
make that happen are a bit ridiculous. Definitely need to add more
control flow words.