Big refactor: getting better but still broken
This fixes some of the brokenness with defining and invoking words
at runtime, but there are still major problems.

I've been working my way through subroutines related to traversing
and extending the dictionary linked lists. I'm re-writing a lot of
that to use the data stack rather than attempting to juggle all the
data in registers. My previous stuff passed everything around in
registers in a manner that was extremely difficult to debug. The
stack based stuff seems a lot easier to follow so far.

I made a couple macros that let me invoke the token handlers with a
wrapper to take care of error checking. This lets me write unrolled
token code pretty conveniently inline with my regular assembly
subroutines. The benefit of the macros is that I can respond to
errors from the bounds checks in the token handlers without having
to make my assembly subroutines completely unreadable due to a
bunch of error checking code.

I also made a `.vars` word to dump variables related to the
dictionary. This is useful for catching bugs that corrupt the
dictionary variables.