Big Refactor: getting less broken
This fixes some segfault and data stack overflow/underflow problems
in my compiler code for defining new words. The main issues were:

1. When I switched addressing from 32-bit pointers in the host CPU
   address space to 16-bit pointers in the virtual CPU address
   space, I missed some spots. Those led to segfaults.

2. The logic for some of my compiling subroutines was a bit muddled
   up. I was losing track of the state of the virtual CPU's data
   stack, and some of the error checks were broken.

It generally seems to be the case that rewriting pure assembly
subroutines to use my pseudo-Forth macros with the virtual data
stack is helpful. When I rewrite buggy subroutines in that manner,
they seem to generally work, and the code is easier to follow.

Lots of the tests are still failing, but not with segfaults, so
that's progress.