Big Refactor: It works! Most tests pass now.
The code for compiling the address of a jump or call target was
dereferencing an already-dereferenced pointer. Naturally, that
messed up nested calls. Fixed it. Seems to work fine now.

There are still a few failing tests, but that mostly has to do with
the fact that I changed some fundamental behavior about how the
outer interpreter, inner interpreter, and compiling words use the
stack. It used to be that I did all that stuff in assembly with
register juggling to avoid touching the data stack. Now, I'm using
the data stack for the interpreters and compiler because it makes
everything much simpler, and it seems more true to how traditional
Forth compilers worked.