Untangle outer loop error handlers
The error handlers for the outer loop were trying to be too sneaky
in a way that could pop one qword too many off the call stack,
leading to a segfault when the markabForth process exits (bye).

This unrolls all the sneaky stuff to make it easier to see that
there are not any `pop rbp` happening. Seems good now.