Add failing tests for compiled var & const
I discovered the answer to my uncertainty about how to handle the
compile branches in `doWord:` for variables and constants. I need
to compile the number, that gets pushed as T when interpreting, as
a literal when compiling.

This commit adds two failing tests that should pass once I've made
the correct change.