Add `if`, `else`, and `then` contitional words
This adds the traditional (weird semantics) Forth IF..ELSE..THEN
words.

I hit a weird segfault in `mStore` that only happened if I added
a bunch of DEBUG and `mDumpVars` calls. I was not able to isolate
a cause. A few of the residual edi vs. rdi changes in this commit
are left over from my searching. This might relate to me using the
SSE registers for data and return stack depths. I'm not sure if
that was a good idea or not. May need to revisit that later.