Rename `IF..ELSE..THEN` to `IF..ELSE..ENDIF`
The old Forth convention of IF..ELSE..THEN is confusing relative to
how popular modern languages do conditionals. I don't want to deal
with trying to explain and justify the weirdness.

There's a reason for why `THEN` comes at the end, but it's
arcane, and could be satisfied just as well by something less
confusing, such as `ENDIF`.  According to Chuck's "Forth - The
Early Years" article (see colorforth website archive, /HOPL.html),
back in 1961, he had to pick something, and that's what he picked.
The ALGOL conditional form with `IF <condition> THEN <true-block>
ELSE <false-block> ...` wouldn't work because Forth needs a word at
the end of the conditional so it can pop the jump address pointer
left by `IF` or `ELSE` and patch the address with the proper target.
Chuck solved the problem by reversing the order of `ELSE` and
`THEN`, and that usage became standardized (but only for Forth).

As far as I can tell, current popular languages seem to do
conditionals in ways that mostly resemble usage from languages
decending from Algol (Simula, C, Pascal, Java, etc.). In the C
tradition, `then` is omitted in favor of parentheses and curly
braces to make blocks. In the Simula and Pascal tradition, `then`
stays, but blocks use `begin` and `end;`.