Add `FOR..NEXT`, fix division ops, update tests
This started off with just adding `FOR` and `NEXT` words to build
counted loops. But, along the way, I found a bug in my `/` and
`mod` words where they would raise a floating point exception if
the dividend was negative.

The main problem was that I didn't use the right size of operand
for the `IDIV` division instruction, so my sign extensions got
messed up.

I fixed the division stuff, added tests to prove it, and updated
the other tests that were using my old recursive looping kludge to
instead use the use the new `FOR..NEXT` stuff.

Example of new looping:
```
: ForA space 5     for   i   . next ;  ForA  5 4 3 2 1 0  OK
: ForB space 5 1 - for   i   . next ;  ForB  4 3 2 1 0  OK
: ForC space 5 1 - for 4 i - . next ;  ForC  0 1 2 3 4  OK
```