Add `1-` word to help set up loop counters
This is a convenience funtion for converting from 1-based counting
to 0-based counting. For example, when you know an array size is 5
elements, and you need to convert that to 0-based indexing, you can
do something like `size 1-`. For example,
```
5 const size  OK
var arr 10 arr ! 11 , 12 , 13 , 14 ,  OK
: .arr size 1- for cr i . ." : " arr i 4 * + ? next ;  OK
.arr
 4:  14
 3:  13
 2:  12
 1:  11
 0:  10  OK
```

I also tweaked the definition of `: , ... ;` because the old way
had the store before its ALLOT. In that case, it happened to work
fine, but that was only because of accidentally getting the benefit
of inner interpreter implementation details that would not apply
outside of compiled words.