Refactor: start to separate VM from Forth system
I've been working loosely from the idea of a virtual stack machine
CPU, memory, and IO -- collectively, the Virtual Machine or VM for
short. But, I haven't been careful to establish clear boundaries
between code for the VM and code for the Forth system that runs on
top of the VM.

This commit is a first step towards pulling out the VM code to be
its own distinct thing. I hope this will make it easier to move
the Forth system from assembly to compiled token code (the Forth
kernel) that can run on top of the VM, and Forth source code
(compiler, number formatting, etc.), that can be compiled at
load-time.

My objective is to significantly reduce the amount of code I will
need to port to get markabForth running on platforms other than
linux-amd64, which is currently the only supported target.