Refactor dictionary data structure + booleans
The main change here is to compile token-code (VM machine code)
inline with the dictionary rather than in a separate memory area.
My old method used code pointers in the dictionary with code in a
different memory block. The only arguable benefit of doing that was
to allow for sneaky recursion stuff with implicit tail calls, like
this:

```
: start do-stuff ( no semicolon!)  : continue do more stuff ;
```

That was useful before I had for-loops working, but now there's no
good reason to keep it. Consolidating so everything goes in the
dictionary simplifies the code considerably in a way that should
make it easier to document the data structure and implementation
details.

Also, I discovered that I got my boolean values swapped from what
normal modern Forths use. Maybe I read some old thing of Chuck's?
Anyway, I flipped the booleans so that true is -1 (non-zero) and
false is 0. The whole boolean thing probably needs a closer look.