Fix `IF`'s conditional logic and update tests
This makes IF work more like the normal pattern for the family of
popular modern programming languages descended from ALGOL. In that
lineage, conditionals normally work like this:
```
if <condition> then <true-block> else <false-block>
```
So, when the condition evaluates to a boolean value of True, the
first block of code runs. When the condition evaluates to False,
the second block of code runs.

For markabForth, after this commit, IF..ELSE..ENDIF now works in a
similar way, like so:
```
<condition> IF <do-when-true> ELSE <do-when-false> ENDIF
```
So, hopefully that behavior will be minimally surprising for folks
trying to read and write markabForth code.

I updated some of the tests to reflect the new behavior of IF.