Allow conditional return from inside `IF` block
This makes it possible to do
```
: a ... IF ... ; ;IF b ;
```
where the call to `b` is conditional, like it was inside an else
block. But, `b` will get tail-call optimized since it comes right
before the final `;` at the end of the word.

The point is to be able to chain calls without filling up the
return stack.

I'm not sure how I feel about the repetition in `; ;if`. Hmm...