Add empty methods for the rest of the opcodes
This is a kind of todo list with an empty VM method for each opcode
that I need to implement for finishing out the instruction set.

To eliminate a bunch of boilerplate repetition, I made a method to
apply a two-argument lambda function to S and T. I'm guessing this
might be slower than unrolling all the stuff. But, in this case, I
would rather have the clarity from less repetition. If this VM is
slow, that's fine. Its purpose is to support development of a ROM
image for the markabForth kernel.