Implement literals and boot vector
This makes it possible to parse integer literals (u8, u16, i32) out
of the stream of instruction tokens and push the ints to the stack.

Of course, to parse numbers out of the token stream, there has to
be a token stream, so I added a memory region for boot code. The
tests for the integer literal tokens work by copying bytes to the
boot area, then stepping the virtual CPU starting from the boot
vector.