Fix automatic stack clearing
If the stack doesn't clear automatically when it gets full, that
can mess up the ability to accept and parse new input from the
outer interpreter. That's really bad, so the stack needs to have
automatic clearing. It's annoying for debugging to lose the stack
contents. But, it's even worse to lose keyboard input.