Implement store and fetch words with lots of tests
New words:
- `!`  and `@`  for  int32 store and fetch
- `w!` and `w@` for uint16 store and fetch
- `b!` and `b@` for uint8  store and fetch

The tests are fairly extensive and verify that range and sign
extension (or the absence thereof) work properly.

The point of having three different sizes of store and fetch is to
provide memory access operations that correspond to how registers
and memory access instructions typically work on a modern-ish
32-bit or 64-bit CPU.

MarkabForth uses 32-bits internally with the intention of good
compatibility across 64-bit computers, 32-bit microcontrollers,
32-bit WebAssembly, and 32-bit soft-CPUs like RV32I on an FPGA.

In the markab VM, registers and stack cells are 32-bits wide, math
operations are 32-bit signed, memory addresses are unsigned uint16,
and bytes are unsigned uint8.