Refactor opcode names for instruction decoder
This is mostly a change to code generation for the opcodes. Many of
the old opcode names were fairly long, and my method for resolving
names to integer values was awkward.

The point of this renaming is to try out some ideas that I think
will be good for building an instruction decoder and assembler.