Revise instruction set for conditional branching
I renamed some of the old instructions to have shorter opcodes and
added two new opcodes for conditional branching:
1. `JZ`: Jump if T is Zero -- this is for implementing conditional
   blocks with `if{ ... }if`
2. `DRJNN`: Decrement R and Jump if Not Negative -- this is to
   implement the conditional backward jump at the end of a for loop
   that is counted by decrementing R down to zero

The tests demonstrate a hand-compiled if-block and for-loop. I've
been experimenting with spellings for the starting and ending
words. At the moment, I kinda like these:
- `if{ ... }if`   (pronounced as "if" and "end-if")
- `for{ ... }for` (pronounced as "for" and "end-for")

The convention I'm considering for the meaning of curly brackets is
that they mark the ends of a conditional block. One end has a jump,
and the other is the target address -- which end is which depends
on the sort of conditional block they belong to.