First draft of `ECALL` peripheral I/O mechanism
I'm gonna try using the RISC-V `ECALL` (Environment CALL) opcode
naming for what linux would refer to as a "syscall".

This commit introduces the mechanic of invoking ECALL with an
argument in T to select which I/O routine is being requested. The
idea is that calling code can do something like `1 ecall` to invoke
a debug log of the data stack contents, `2 ecall` to log the return
stack contents, `5 ecall` to push a copy of the PC register,
`6 ecall` to do a non-blocking read from standard input, etc.

The ecall routines are not fully implemented yet. I just changed
some of the surrounding plumbing, mostly to prepare for being able
to invoke debug logging of stack contents from object code rather
than having to call `VM._log_ds()` or `VM._log_rs()`.

The experiment here is to see if this syscall-style IO mechanic is
better than using hardware style Control Status Registers (CSRs)
with memory mapped addresses. That method is natural on hardware,
but for a VM, maybe it's just extra work with minimal benefits?
Specifying an interface based on ECALL does not rule out use of
memory mapped peripheral CSRs, it just leaves the details up to the
VM implementer.

For example, a Verilog implementation of ECALL could use an array
of function pointers ("vectors" in CPU lingo) to call ROM routines
that access memory mapped CSRs. A software implementation could
skip all that and just validate the arguments then pass them on to
an appropriate syscall from the host OS.