Add `RDROP` opcode for end of for loop
RDROP drops R from the top of the return stack, which is needed at
the end of a counted loop. The previous method was to move R to the
data stack, then drop from the data stack. Dropping R is going to
be a common thing, so it might as well get its own opcode.