Modify `BZ` opcode to drop T
Previously I was leaving the boolean conditional value, T, on top
of the data stack for BZ as used in `if{ ... }if`. That meant a lot
of dropping inside of and after if-blocks. For my test cases so
so far, this change will make the machine code shorter. On the
other hand, if I want to implement something like a case statement,
with multiple comparisons against the same value, including the
drop in BZ may be undesirable. For now, I'll do it this way.