First draft of markab_kernel.mkb
This is the start of source code for a markab kernel rom image.
When finished, it will include an outer interpreter, inner
interpreter, assembler, and compiler for markab, written in markab.
My plan is to bootstrap the kernel rom with a markab assembler and
compiler written in Python (which does not exist yet).