Work toward getting compiler up and running
I'm trying to figure out how I want compiling and assembling to
behave. In my previous experiment in `../repl`, I just wrote the
assembler code to put raw bytes into the dictionary entries without
bothering to write Forth-like words abstracting what was happening.

Now I'm trying to work out a reasonable way of writing it all down
with higher level abstractions in Markab. One of the tricky things
is that I want to have a dictionary item type just for opcodes.
This will let my inner interpreter either run the opcode as code in
interpret mode or compile it with inline optimization in compile
mode.

I also added a way to track the nesting level of if{ and for{
blocks. Tracking nesting will let ; decide whether or not it needs
to exit compile mode. A ; inside an if{ or for{ should leave
compile mode set, but at the end of a word, it should clear compile
mode.