Implement inner interpreter and add `CALL` opcode
This implements compile-mode and interpret-mode behavior for the
inner interpreter.

I realized that I needed an opcode to make a subroutine call to an
address from T. Since `JAL` takes its address from the instruction
stream, it only works for compiled calls. I added a new opcode,
`CALL`, to call an address from T so the inner interpreter can do
what it needs to do in interpret-mode.