First draft of real bootstrap rom compiler
This compiles kernel.mkb into kernel.rom, which markab_vm.py can
actually run. But, it's very buggy and useless at the moment. It
gets to the point of reading a line of input, but it gets confused
as soon as it attempts to parse that line. I can tell that it's
having trouble in `word` by watching a trace of the VM's program
counter register.

This is not at all surprising considering that I wrote the whole
kernel speculatively without any way to do incremental testing.
After some debugging earlier today, I got all that code to at
least compile, but it needs a lot more testing and debugging.