Get `readline` and `word` working with debug trace
This makes the kernel substantially less broken. The REPL is now
capable of responding to some simple stuff like `iod`, `true`, and
`swap`. But, number parsing does not work.

There was a nasty bug in the VM where `JAL` was overwriting `R`
with `PC`, in addition to pushing `PC`. I fixed that so subroutine
calls work better. That was one of the problems with `readline` and
`word` not working.

I also added debug tracing to the VM that is controlled by a
`DEBUG` flag which is kinda buried in the code. The tracing uses a
symbol table and opcode lookup to show the program counter, word,
opcode, and data stack for each step of the clock. It's nice.