Refactor OK/ERR prompt between VM and tests
I started off with OK/ERR prompts built into the VM's routines for
printing the data stack and return stack. That was good to save
some code while writing tests for VM functions and object code
instruction decoding.

But, now that I have a kernel, it needs to be able to manage its
own OK/ERR prompts. So, this removes the automatic prompting from
the VM and replaces its role in the tests with manual calls to
generate the prompts. This fixes the double-prompting that was
happening in the kernel shell.