Combine decimal and hex number parsing
The size and complexity of the number parsing code was bugging me.
There was a lot of duplication due to my desire to support
negative decimal numbers with a leading '-'.

Generalizing the behavior of a leading '-' to also negate hex
numbers let me drastically simplify the number parsing. The idea
of a negative hex number seems odd, but allowing that is much
better than telling people they have to negate their own decimal
numbers with annoying workarounds like `3 inv 1+` or `0 3 -`.