Reproduce and fix weird bug, implement fence
The problem with the weird bug glitching output when `dowords`
called `spaces` was because of an extra unmatched `}if` in the
preceeding `@a_atoi?` word that was confusing the bootstrap
compiler.

The bootstrap compiler tracks nesting levels of `if` and `for` to
facilitate tail call optimization and returning from inside loops.
But, it wasn't checking to see if the nesting level counter ever
went negative. Naturally, it was going negative, and that broke the
mechanism for deciding whether int literals should get pushed to
the stack or compiled.

The arguments at the end of `dowords` were getting pushed at
compile time, rather than compiled as literals. So, the call to
`spaces` was messing up the stack and causing chaos since the
literal for its argument was not compiled into the code.

I fixed `@a_atoi?` and added a negative nesting check to the
bootstrap complier, so this should not happen again.

Along the way, I also implemented the VM `fence` low-memory write
protect mechanism that I've been thinking about for a while.
Writing to the VM's fence register makes it so that the storing
opcodes can only write to memory above the fence address. I
thought maybe the dictionary was getting corrupted, but the problem
was happening earlier in the bootstrap compiler.