Add `ERR`, `CLERR` and finish `FENCE`, with tests
This sets the stage for the kernel's outer interpreter to manage
its own error messages related to memory access faults, among other
error conditions. Previously, the VM was hardcoded to print error
messages and, in some cases, interrupt control flow.

I implemented the fence checks for storing words so that stack side
effects (drop data and address) happen regardless of whether the
store was successful. The general idea is to let error checking be
an opt-in thing for kernel code rather than hijacking control flow
in ways that make it hard for the kernel to detect and recover from
error conditions.

I'm not sure if this is the best approach. It might be better, for
example, to have an error vector that PC gets set to after an
error. Then, whatever code lives at the error vector address can
try to make sense of what happened. Not sure.

`ERR` pushes the error register to the data stack, and `CLERR`
clears the error register.