Rework for-loops using new debug and test stuff
This started with my attempt to define and run a word that included
a for loop, but it quickly ballooned into an epic quest.

To debug the weird for loop bug, I ended up making several changes:
- Instead of a 2 instruction sequence to end for loops, there is
  now 1 new instruction: `BFOR`, standing for branch-to-for
- New `TRON` (trace on) and `TROFF` (trace off) instructions allow
  fine grained control of when trace log messages are generated
- New `IODUMP` instruction prints hexdumps of memory ranges
- New `IODOT` instruction prints number on top of stack
- Makefile's `make test` now runs an additional test sequence to
  test the kernel rom's ability to compile and run Markab code.
  (the older tests aim at testing the VM with object code)

Part of the problem with the for loop bug, which I eventually
fixed, was that the number compiling code was leaving extra stuff
on the stack. That messed up the `BFOR` target address.