Speed up `find`, `strcmp`, and dictionary: -290ms
This gets the rom test running about 290ms faster. In decreasing
order of impact, changes include:
1. Optimize the Markab code for `strcmp` and `find` in the kernel
   to use fewer instructions. This includes eliminating the opcodes
   for `>x`, `x`, `>y`, and `y`. (saves ~230ms)
2. Sort opcodes in the code generation list by approximate
   frequency of use so dictionary searches complete faster (~40ms)
3. Fix a couple more 16-bit little-endian conversions (~20ms)

Based on what I see here, I expect it would be very much worth my
trouble to implement a hashmap for the dictionary. As it stands,
every number that I compile does a worst-case traversal of the full
linked list, and many of the other words do a lot of searching.

Rewriting the VM in C or assembler, I could probably get away with
more dictionary searching. But, as a general principle, I want to
avoid input latency and CPU spikes due to poor algorithm and data
structure choices. For now, Python being slow is good motivation to
make sure my kernel code is efficient.