Change `JMP` and `JAL` to PC-relative addressing
I think this is the last change I need so that compiled object code
becomes fully relocatable. The `CALL` instruction can and should
use absolute addressing because it takes its address argument on
the stack as the result of an interactive dictionary search. In
that case, the rom loading or compilation will have already
happened, so there is no need to worry about code being relocated.