Get rid of `MRT` / `r>` as it is not needed
This was left over from some of my earliest opcode stuff, before I
figured out how to do good for loops. The combination of `R` and
`RDROP`, which I definitely need, could substitute for `MRT`. But,
so far I haven't needed that. Mainly I use the `rdrop ;` combo to
exit from if-blocks inside of for-loops.