Improve string hash function searching
This adds a new polynomial string hash parameter searching thing,
written in C. I also improved the slower Python one.

The point of this stuff is to figure out what kind of tradeoffs are
possible with hashmap table size vs. worst-case number of links to
follow for the linked list in each bin. I can probably get the list
length down to 2 or 3 items if I use 94 or 95 bins. So, with a
16-bit address field for the list head of each bin, that would add
about 190 bytes to the dictionary. That seems probably fine.