Compute a pretty good hash function for 64 bins
This set of parameters and bin size gives a median of 3 links to
follow per lookup, with a worst case of 4 links to follow. That's
pretty good, really. If I were to up the bin count to somewhere in
the range of 100-128 bins, I could only bring that down to a median
of 2 links to follow per bin with a worst case of 3 links. Probably
not worth the extra dictionary bytes to use a larger hash table.