Convert polynomial hash back to shift and mask
It turns out that I can get a pretty good result with 64 bins and
a coefficient of 2 (left shift by 1). So, I converted the hash
functions back to only using shift and mask operations. This
sacrificies a little performance on fancy CPUs with fast multiply
and divide in exchange for better potential performance on embedded
processors with slow multiply and divide.