Add string hash parameters to core vocab
This adds constants for parameters of the polynomial string hashing
function to the core vocabulary. I also recomputed the parameters
to include the dictionary entries for their names.

The point of this is to get ready for writing Markab code to modify
the `find` and `create` kernel words to use a hashmap.