Implement hashmaps for core and extensible vocab!
This is the big hashmap thing I've been working towards for a while
now. It seems to be pretty much what I was hoping for!

Previously, I had one dictionary which was a linked list. This
change splits the dictionary into a core vocabulary and extensible
vocabulary. Each vocabulary gets its own 64 bin hashmap with a
short linked list for each bin. The core vocabulary's hashmap is
compiled into the kernel rom then copied to ram by the VM. The
extensible vocabulary's hashmap is created at runtime.

According to my cProfile stuff, this change cuts the runtime for
`make romtest` (test of compiling inside the VM) down from about
610ms to around 220ms. So, the delay is barely noticeable now. When
I tried reconfiguring the hashmap for 128 bins, it actually made
the romtest a little slower. Probably that was due to the need for
the boot code to initialize an extra 128 bytes of hashmap pointers
for the extensible vocabulary.